integration_services/parser_service/requirements.txt
fastapi
uvicorn[standard]
lxml
google-cloud-pubsub
google-cloud-storage